Se agregan modificaciones en cart

+++ routes/products.routes.js
@@ -3,10 +3,11 @@ const express = require("express");
 const routerProducts = express.Router();
 
 const ContainerProducts = require("../containers/products.js");
+const ContainerAdmin = require("../containers/security.js");
 
 /* -------------------------------- Instancia de Express ------------------------ */
 const productsApi = new ContainerProducts();
-
+const roleAdminApi = new ContainerAdmin();
 /* -------------------------------- Rutas -------------------------------- */
 // https://www.iconfinder.com/free_icons
 /*
@@ -40,20 +41,24 @@ routerProducts.get("/:id?", (req, res) => {
 //agrega un producto y usamos middlewares
 routerProducts.post("/", (req, res) => {
   try {
-    const { name, description, code, thumbnail, price, stock } = req.body;
+    if (!roleAdminApi.roleAdmin()) {
+      res.status(401).json(roleAdminApi.notAutorized("POST", "/products"));
+    } else {
+      const { name, description, code, thumbnail, price, stock } = req.body;
+ 
+      let producto = {
+        timestamp: Date.now(),
+        name: name,
+        description: description,
+        code: code,
+        thumbnail: thumbnail,
+        price: price,
+        stock: stock,
+      };
 
-    let producto = {
-      timestamp: Date.now(),
-      name: name,
-      description: description,
-      code: code,
-      thumbnail: thumbnail,
-      price: price,
-      stock: stock
-    };
-
-    const response = productsApi.save(producto);
-    res.status(200).json(response);
+      const response = productsApi.save(producto);
+      res.status(200).json(response);
+    }
   } catch (error) {
     res.status(404).json({ msg: `Error al agregar Producto: ${error}` });
   }
@@ -64,21 +69,25 @@ routerProducts.post("/", (req, res) => {
 //recibe y actualiza un producto según su id
 routerProducts.put("/:id", (req, res) => {
   try {
-    const { id } = req.params;
-    const { name, description, code, thumbnail, price, stock } = req.body;
+    if (!roleAdminApi.roleAdmin()) {
+      res.status(401).json(roleAdminApi.notAutorized("PUT", "/products"));
+    } else {
+      const { id } = req.params;
+      const { name, description, code, thumbnail, price, stock } = req.body;
 
-    let producto = {
-      timestamp: Date.now(),
-      name: name,
-      description: description,
-      code: code,
-      thumbnail: thumbnail,
-      price: price,
-      stock: stock
-    };
+      let producto = {
+        timestamp: Date.now(),
+        name: name,
+        description: description,
+        code: code,
+        thumbnail: thumbnail,
+        price: price,
+        stock: stock,
+      };
 
-    const response = productsApi.update(producto, id);
-    res.status(200).json(response);
+      const response = productsApi.update(producto, id);
+      res.status(200).json(response);
+    }
   } catch (error) {
     res.status(404).json({ msg: `Error al actualizar Producto: ${error}` });
   }
@@ -87,14 +96,17 @@ routerProducts.put("/:id", (req, res) => {
 //elimina un producto según su id
 routerProducts.delete("/:id", (req, res) => {
   try {
-    const { id } = req.params;
+    if (!roleAdminApi.roleAdmin()) {
+      res.status(401).json(roleAdminApi.notAutorized("DELETE", "/products"));
+    } else {
+      const { id } = req.params;
 
-    const response = productsApi.delete(id);
-    res.status(200).json(response);
+      const response = productsApi.delete(id);
+      res.status(200).json(response);
+    }
   } catch (error) {
     res.status(404).json({ msg: `Error al eliminar Producto: ${error}` });
   }
 });
-
 
 module.exports = routerProducts;
